+++ src/components/render.js
@@ -1,6 +1,6 @@
 import { $newTodoTitle, $todoList, $todoCount, $filters } from './dom.js';
 import todoList from '../components/todoList.js';
-import getCount from '../components/getCount.js';
+import getCount from '../utils/getCount.js';
 import getFilter from '../components/getFilter.js';
 const render = ({ todoItem, option }) => {
   $todoList.innerHTML = todoList(todoItem, option);
